+++ src/test/compile-fail/issue-6702.rs
@@ -9,11 +9,11 @@
 // except according to those terms.
 
 struct Monster {
-    damage: int
+    damage: isize
 }
 
 
 fn main() {
     let _m = Monster(); //~ ERROR `Monster` is a structure name, but
-    //~^ NOTE Did you mean to write: `Monster { /* fields */ }`?
+    //~^ HELP Did you mean to write: `Monster { /* fields */ }`?
 }
