+++ src/test/compile-fail/tail-typeck.rs
@@ -10,8 +10,8 @@
 
 // error-pattern: mismatched types
 
-fn f() -> int { return g(); }
+fn f() -> isize { return g(); }
 
-fn g() -> uint { return 0u; }
+fn g() -> usize { return 0; }
 
 fn main() { let y = f(); }
